fix invalid data type

+++ features/transaction/checkout/src/main/java/com/tokopedia/checkout/view/uimodel/CrossSellModel.kt
@@ -8,7 +8,7 @@ data class CrossSellModel(
         var id: String = "",
         var checkboxDisabled: Boolean = false,
         var isChecked: Boolean = false,
-        var price: Long = 0,
+        var price: Double = 0.0,
         var info: CrossSellInfoModel = CrossSellInfoModel(),
         var orderSummary: CrossSellOrderSummaryModel = CrossSellOrderSummaryModel(),
         var bottomSheet: CrossSellBottomSheetModel = CrossSellBottomSheetModel(),
